requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
jinja2
torch
torchvision
Pillow
numpy